+++ src/MonadIO.ts
@@ -3,15 +3,15 @@
  *
  * @since 2.0.0
  */
-import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3, URIS4, Kind4 } from './HKT'
+import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
 import { IO } from './IO'
-import { Monad, Monad1, Monad2, Monad3, Monad2C, Monad3C, Monad4 } from './Monad'
 
 /**
  * @category type classes
  * @since 2.0.0
  */
-export interface MonadIO<M> extends Monad<M> {
+export interface MonadIO<M> {
+  readonly URI: M
   readonly fromIO: <A>(fa: IO<A>) => HKT<M, A>
 }
 
@@ -19,7 +19,8 @@ export interface MonadIO<M> extends Monad<M> {
  * @category type classes
  * @since 2.0.0
  */
-export interface MonadIO1<M extends URIS> extends Monad1<M> {
+export interface MonadIO1<M extends URIS> {
+  readonly URI: M
   readonly fromIO: <A>(fa: IO<A>) => Kind<M, A>
 }
 
@@ -27,7 +28,8 @@ export interface MonadIO1<M extends URIS> extends Monad1<M> {
  * @category type classes
  * @since 2.0.0
  */
-export interface MonadIO2<M extends URIS2> extends Monad2<M> {
+export interface MonadIO2<M extends URIS2> {
+  readonly URI: M
   readonly fromIO: <E, A>(fa: IO<A>) => Kind2<M, E, A>
 }
 
@@ -35,7 +37,9 @@ export interface MonadIO2<M extends URIS2> extends Monad2<M> {
  * @category type classes
  * @since 2.0.0
  */
-export interface MonadIO2C<M extends URIS2, E> extends Monad2C<M, E> {
+export interface MonadIO2C<M extends URIS2, E> {
+  readonly URI: M
+  readonly _E: E
   readonly fromIO: <A>(fa: IO<A>) => Kind2<M, E, A>
 }
 
@@ -43,7 +47,8 @@ export interface MonadIO2C<M extends URIS2, E> extends Monad2C<M, E> {
  * @category type classes
  * @since 2.0.0
  */
-export interface MonadIO3<M extends URIS3> extends Monad3<M> {
+export interface MonadIO3<M extends URIS3> {
+  readonly URI: M
   readonly fromIO: <R, E, A>(fa: IO<A>) => Kind3<M, R, E, A>
 }
 
@@ -51,7 +56,9 @@ export interface MonadIO3<M extends URIS3> extends Monad3<M> {
  * @category type classes
  * @since 2.2.0
  */
-export interface MonadIO3C<M extends URIS3, E> extends Monad3C<M, E> {
+export interface MonadIO3C<M extends URIS3, E> {
+  readonly URI: M
+  readonly _E: E
   readonly fromIO: <R, A>(fa: IO<A>) => Kind3<M, R, E, A>
 }
 
@@ -59,6 +66,7 @@ export interface MonadIO3C<M extends URIS3, E> extends Monad3C<M, E> {
  * @category type classes
  * @since 2.4.4
  */
-export interface MonadIO4<M extends URIS4> extends Monad4<M> {
+export interface MonadIO4<M extends URIS4> {
+  readonly URI: M
   readonly fromIO: <S, R, E, A>(fa: IO<A>) => Kind4<M, S, R, E, A>
 }