+++ CHANGELOG.md
@@ -48,6 +48,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `intercalate` pipeable (@gcanti)
   - `IOEither`
     - remove `getIOValidation`, use `getApplicativeIOValidation` and `getAltIOValidation` instead (@gcanti)
+  - `MonadIO`
+    - drop `Monad` dependency (@gcanti)
   - `Monoid`
     - swap execution order in `getEndomorphismMonoid` (@gcanti)
   - `ReaderEither`